Made allowing sender ID spoofing configurable.

+++ vacuumworld/__init__.py
@@ -5,6 +5,7 @@ from sys import version_info
 
 from .config_manager import ConfigManager
 from .model.actions.vwactions import VWAction
+from .model.actions.vwactions import VWCommunicativeAction
 from .model.actions.effort import ActionEffort
 from .model.actor.actor_mind_surrogate import ActorMindSurrogate
 from .model.actor.user_mind_surrogate import UserMindSurrogate
@@ -34,29 +35,34 @@ def python_version_check() -> None:
         exit(1)
 
 
-def version_check() -> None:
-    config: dict = ConfigManager(config_file_path=CONFIG_FILE_PATH).load_config()
+def version_check(config: dict) -> None:
     version_number: str = config["version_number"]
 
     print("VacuumWorld version: {}.\n".format(version_number))
 
     call(["wget", "https://raw.githubusercontent.com/dicelab-rhul/vacuumworld/main/vacuumworld/{}".format(CONFIG_FILE_NAME)], stdout=DEVNULL, stderr=DEVNULL)
 
-    remote_config: dict = ConfigManager(config_file_path=CONFIG_FILE_NAME).load_config()
-    latest_version_number: str = remote_config["version_number"]
+    try:
+        remote_config: dict = ConfigManager(config_file_path=CONFIG_FILE_NAME).load_config()
+        latest_version_number: str = remote_config["version_number"]
 
-    if version_number != latest_version_number:
-        print("WARNING: Your version of VacuumWorld is outdated. The latest version is {}.".format(latest_version_number))
-        print("Please update VacuumWorld by running `./update_vw.sh` from a terminal, after navigating to the cloned directory.\n")
-    else:
-        print("Your version of VacuumWorld is up-to-date.\n")
-
-    os.remove(CONFIG_FILE_NAME)
+        if version_number != latest_version_number:
+            print("WARNING: Your version of VacuumWorld is outdated. The latest version is {}.".format(latest_version_number))
+            print("Please update VacuumWorld by running `./update_vw.sh` from a terminal, after navigating to the cloned directory.\n")
+        else:
+            print("Your version of VacuumWorld is up-to-date.\n")
+    finally:
+        if os.path.exists(CONFIG_FILE_NAME):
+            os.remove(CONFIG_FILE_NAME)
 
 
 def run(default_mind=None, white_mind=None, green_mind=None, orange_mind=None, **kwargs) -> None:
     python_version_check()
-    version_check()
+
+    config: dict = ConfigManager(config_file_path=CONFIG_FILE_PATH).load_config()
+    VWCommunicativeAction.SENDER_ID_SPOOFING_ALLOWED = config["sender_id_spoofing_allowed"]
+
+    version_check(config=config)
 
     # Safeguard against crashes on Windows and every other OS without SIGTSTP.
     if hasattr(signal, "SIGTSTP"):
@@ -67,9 +73,6 @@ def run(default_mind=None, white_mind=None, green_mind=None, orange_mind=None, *
     __assign_efforts_to_actions(**kwargs)
 
     white_mind, green_mind, orange_mind = __process_minds(default_mind, white_mind, green_mind, orange_mind)
-
-    config: dict = ConfigManager(config_file_path=CONFIG_FILE_PATH).load_config()
-
     user_mind: UserMindSurrogate = UserMindSurrogate(difficulty_level=UserDifficulty(config["default_user_mind_level"]))
 
     if "gui" in kwargs and not kwargs.get("gui"):
